refactor(rate-limit): merge rate-limiters.ts into rate-limit.ts
Two files for 60 lines of code where rate-limiters.ts existed only to
hold module-level singletons of RateLimiter. Single bidirectional
dependency. Per the improve-codebase-architecture audit, the split is
accidental — the class and its named instances are one concept.
Callers update their import path from ./rate-limiters to ./rate-limit.
